fix: updating camera modes on input

+++ frontend/src/app/components/upload/upload.component.ts
@@ -21,6 +21,7 @@ export class UploadComponent implements OnInit {
   private sceneObjects: { geometry: BufferGeometry, mesh: Mesh }[] = [];
   public controlOptions: any[] = [];
   loading = false;
+  keepGrid = false;
 
   grid: GridHelper;
   backgroundColor: string = '#ffffff';
@@ -75,7 +76,7 @@ export class UploadComponent implements OnInit {
   }
 
   takeSnapshot() {
-    this.grid.visible = false;
+    if (!this.keepGrid) this.grid.visible = false;
     this.sceneController.hideControls();
     const image = this.engineService.takeSnapshot();
     this.snapshotEncoded = image;
@@ -97,7 +98,7 @@ export class UploadComponent implements OnInit {
     const content = this.postContent;
 
     this.sceneController.removeControls();
-    this.engineService.removeFromScene(this.grid);
+    if (!this.keepGrid) this.engineService.removeFromScene(this.grid);
 
     const scene = this.engineService.saveAndDestroy();
     const blob = new Blob([JSON.stringify(scene)], { type: "application/json;charset=utf-8" })
@@ -129,12 +130,7 @@ export class UploadComponent implements OnInit {
 
   updateColor(upload: Upload) {
     if (!upload.mesh) return;
-    if (typeof (upload.color as any) === 'string') {
-      (upload.mesh.material as MeshStandardMaterial).color.setHex( Number((upload.color as any).replace('#', '0x')));
-    } else {
-      (upload.mesh.material as MeshStandardMaterial).color.setHex(upload.color);
-    }
-
+    (upload.mesh.material as MeshStandardMaterial).color.setHex( Number(upload.color.replace('#', '0x')));
   }
 
   centerUpload(upload: Upload) {
@@ -181,7 +177,7 @@ export class UploadComponent implements OnInit {
           controls,
           snapshot: this.engineService.createMeshSnapshot(mesh, this.sceneObjects.map(c => c.mesh)),
           snapshotImage: null,
-          color: 0xffffff,
+          color: '0xffffff',
         };
 
         this.uploads.push(upload);
@@ -200,5 +196,9 @@ export class UploadComponent implements OnInit {
       }
     };
     reader.readAsText(files[0]);
+  }
+
+  changeTab(event: any) {
+    this.sceneController.updateKeyListener(event.index != 0);
   }
 }
